# TM8/tm8-game-mobile-release-0.6.0/input_folder/dev/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/users/{userId}/games": {
            "get": {
                "operationId": "UserController_getUserGames",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "description": "User ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGamesResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User"
                ]
            }
        },
...
        "/api/v1/chat/channel": {
            "post": {
                "operationId": "ChatController_createChannel",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateChannelInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateChannelResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Chats"
                ]
            }
        },
        "/api/v1/chat/user/channels": {
            "get": {
                "operationId": "ChatController_listUsers",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Page size",
                        "schema": {
                            "minimum": 1,
                            "maximum": 50,
                            "default": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "username",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatChannelPaginatedResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Chats"
                ]
            }
        },
        "/api/v1/chat/token/refresh": {
            "post": {
                "operationId": "ChatController_refreshChatToken",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatRefreshTokenResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatRefreshTokenResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Chats"
                ]
            }
        },
...
        "/api/v1/landing-page/contact-form": {
            "post": {
                "operationId": "LandingPageController_sendContactEmail",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactFormInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Landing Page"
                ]
            }
        },
        "/api/v1/games": {
            "get": {
                "operationId": "GameController_getUserReportTypes",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GameResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Game"
                ]
            }
        },
        "/api/v1/games/{game}/preference-form": {
            "get": {
                "operationId": "GameController_getPreferenceForm",
                "parameters": [
                    {
                        "name": "game",
                        "required": true,
                        "in": "path",
                        "description": "Game name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GamePreferenceInputResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Game"
                ]
            }
        }
    },
    "info": {
        "title": "TM8 API",
        "description": "TM8 API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "PaginationMetaResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "number",
                        "description": "Page number"
                    },
                    "limit": {
                        "type": "number",
                        "description": "Page size"
                    },
                    "itemCount": {
                        "type": "number",
                        "description": "Total items without pagination"
                    },
                    "pageCount": {
                        "type": "number",
                        "description": "Total pages with current page size"
                    },
                    "hasPreviousPage": {
                        "type": "boolean",
                        "description": "Flag indicating if there is a previous page"
                    },
                    "hasNextPage": {
                        "type": "boolean",
                        "description": "Flag indicating if there is a next page"
                    }
                },
                "required": [
                    "page",
                    "limit",
                    "itemCount",
                    "pageCount",
                    "hasPreviousPage",
                    "hasNextPage"
                ]
            },
...
            "TGMDExceptionResponse": {
                "type": "object",
                "properties": {
                    "origin": {
                        "enum": [
                            "HTTP",
                            "Mongo",
                            "TGMD",
                            "Unknown"
                        ],
                        "type": "string"
                    },
                    "status": {
                        "type": "number",
                        "description": "HTTP-compatible status code"
                    },
                    "code": {
                        "type": "string"
                    },
                    "exception": {
                        "type": "string",
                        "description": "Exception name"
                    },
                    "detail": {
                        "type": "string",
                        "description": "Human-readable error message"
                    }
                },
                "required": [
                    "origin",
                    "status",
                    "code",
                    "exception",
                    "detail"
                ]
            },
...
            "UserGameResponse": {
                "type": "object",
                "properties": {
                    "displayName": {
                        "type": "string"
                    },
                    "game": {
                        "enum": [
                            "call-of-duty",
                            "fortnite",
                            "apex-legends",
                            "rocket-league"
                        ],
                        "type": "string"
                    }
                },
                "required": [
                    "displayName",
                    "game"
                ]
            },
            "UserGamesResponse": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "games": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserGameResponse"
                        }
                    }
                },
                "required": [
                    "userId",
                    "games"
                ]
            },
...
            "CreateChannelInput": {
                "type": "object",
                "properties": {
                    "members": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "members"
                ]
            },
            "CreateChannelResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "ChatChattingWithResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "online": {
                        "type": "boolean"
                    },
                    "avatar": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "username",
                    "online"
                ]
            },
            "ChatChannelResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "chattingWith": {
                        "$ref": "#/components/schemas/ChatChattingWithResponse"
                    },
                    "lastMessage": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "chattingWith",
                    "lastMessage"
                ]
            },
            "ChatChannelPaginatedResponse": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChatChannelResponse"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMetaResponse"
                    }
                },
                "required": [
                    "items",
                    "meta"
                ]
            },
            "ChatRefreshTokenResponse": {
                "type": "object",
                "properties": {
                    "chatToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "chatToken"
                ]
            },
...
            "ContactFormInput": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "email",
                    "message"
                ]
            },
            "GameResponse": {
                "type": "object",
                "properties": {
                    "display": {
                        "type": "string"
                    },
                    "game": {
                        "type": "string",
                        "enum": [
                            "call-of-duty",
                            "fortnite",
                            "apex-legends",
                            "rocket-league"
                        ]
                    },
                    "iconKey": {
                        "type": "string"
                    }
                },
                "required": [
                    "display",
                    "game",
                    "iconKey"
                ]
            },
            "SliderOptionResponse": {
                "type": "object",
                "properties": {
                    "minValue": {
                        "type": "string"
                    },
                    "maxValue": {
                        "type": "string"
                    },
                    "attribute": {
                        "enum": [
                            "arena",
                            "duo",
                            "intermediate",
                            "aggressive",
                            "team-player",
                            "find-people",
                            "center-circle",
                            "casual",
                            "hoops",
                            "2v2",
                            "beginner",
                            "playing-level",
                            "how-you-play",
                            "mixtape",
                            "control",
                            "battle-royale",
                            "battle-royale-build",
                            "battle-royale-no-build",
                            "advanced",
                            "creative-mode",
                            "trio",
                            "squad",
                            "first-time",
                            "competitive",
                            "pro",
                            "legend",
                            "chase-kill",
                            "zone-walk",
                            "resurgence",
                            "plunder",
                            "gun-game",
                            "ranked",
                            "tournaments",
                            "snow-day",
                            "dropshot",
                            "rumble",
                            "3v3",
                            "4v4",
                            "training",
                            "firing-range",
                            "ranked-league",
                            "assault",
                            "skirmisher",
                            "recon",
                            "controller",
                            "support",
                            "team-deathmatch",
                            "run-gun",
                            "scorer",
                            "offensive",
                            "extras",
                            "firing-squad",
                            "demo-heavy",
                            "passer",
                            "anchor",
                            "1v1",
                            "offensive-scorer",
                            "offensive-passer",
                            "mid-scorer",
                            "mid-defender",
                            "defensive-scorer",
                            "defensive-passer",
                            "warzone",
                            "quick-games",
                            "starting-timestamp",
                            "ending-timestamp"
                        ],
                        "type": "string"
                    }
                },
                "required": [
                    "minValue",
                    "maxValue",
                    "attribute"
                ]
            },
            "DropdownOptionResponse": {
                "type": "object",
                "properties": {
                    "display": {
                        "type": "string"
                    },
                    "attribute": {
                        "type": "string"
                    }
                },
                "required": [
                    "display",
                    "attribute"
                ]
            },
            "GamePreferenceInputResponse": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "enum": [
                            "select",
                            "multi-select",
                            "slider",
                            "dropdown"
                        ],
                        "type": "string"
                    },
                    "sliderOptions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SliderOptionResponse"
                        }
                    },
                    "selectOptions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SelectOptionResponse"
                        }
                    },
                    "dropdownOptions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DropdownOptionResponse"
                        }
                    }
                },
                "required": [
                    "title",
                    "type"
                ]
            },
            "SelectOptionResponse": {
                "type": "object",
                "properties": {
                    "display": {
                        "type": "string"
                    },
                    "attribute": {
                        "enum": [
                            "arena",
                            "duo",
                            "intermediate",
                            "aggressive",
                            "team-player",
                            "find-people",
                            "center-circle",
                            "casual",
                            "hoops",
                            "2v2",
                            "beginner",
                            "playing-level",
                            "how-you-play",
                            "mixtape",
                            "control",
                            "battle-royale",
                            "battle-royale-build",
                            "battle-royale-no-build",
                            "advanced",
                            "creative-mode",
                            "trio",
                            "squad",
                            "first-time",
                            "competitive",
                            "pro",
                            "legend",
                            "chase-kill",
                            "zone-walk",
                            "resurgence",
                            "plunder",
                            "gun-game",
                            "ranked",
                            "tournaments",
                            "snow-day",
                            "dropshot",
                            "rumble",
                            "3v3",
                            "4v4",
                            "training",
                            "firing-range",
                            "ranked-league",
                            "assault",
                            "skirmisher",
                            "recon",
                            "controller",
                            "support",
                            "team-deathmatch",
                            "run-gun",
                            "scorer",
                            "offensive",
                            "extras",
                            "firing-squad",
                            "demo-heavy",
                            "passer",
                            "anchor",
                            "1v1",
                            "offensive-scorer",
                            "offensive-passer",
                            "mid-scorer",
                            "mid-defender",
                            "defensive-scorer",
                            "defensive-passer",
                            "warzone",
                            "quick-games",
                            "starting-timestamp",
                            "ending-timestamp"
                        ],
                        "type": "string"
                    },
                    "cascade": {
                        "$ref": "#/components/schemas/GamePreferenceInputResponse"
                    }
                },
                "required": [
                    "display",
                    "attribute"
                ]
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}